backend/requirements.txt
fastapi
uvicorn[standard]
supabase-py
python-dotenv
httpx
pydantic
